Project #1-Simple ML
Built a simple gender classifier by training various models using a given data set. The models based their predictions on whether or not the given measurements [height (cm), weight (kg), shoe size (European sizing)] belonged to a male or female person. I found the Gaussian Naive Bayes classifier to be the most accurate with its predictions for this data set, and the KNeighbors classifier to be the worst predictor of gender.